move school id fetch method within school score data service context
