+++ functions-app/calculate-score-v2/schools-scores.data.service.js
@@ -1,3 +1,4 @@
+const R = require('ramda')
 const sqlService = require('../lib/sql/sql.service')
 const { TYPES } = sqlService
 
@@ -22,4 +23,15 @@ module.exports.sqlExecuteGetSchoolScores = async (checkWindowId, schoolId) => {
     }
   ]
   return sqlService.query(sql, params)
+}
+
+/**
+ * Get all school ids
+ * @return {Promise<object>}
+ */
+module.exports.sqlFindSchoolIds = async () => {
+  const sql = `SELECT id FROM [mtc_admin].[school]`
+  const params = []
+  const result = await sqlService.query(sql, params)
+  return R.map(i => i.id, result)
 }
