+++ functions-app/calculate-score-v2/v1.spec.js
@@ -6,7 +6,6 @@ const moment = require('moment')
 const config = require('../config')
 const schoolScoresDataService = require('./schools-scores.data.service')
 const checkWindowDataService = require('./check-window.data.service')
-const schoolDataService = require('./school.data.service')
 const pupilResultsDiagnosticCache = require('./pupil-results-diagnostic-cache.data.service')
 const redisCacheService = require('../lib/redis-cache.service')
 const v1 = require('./v1')
@@ -16,7 +15,7 @@ describe('calculate-score-v2: v1', () => {
   describe('process', () => {
     it('executes get school scores store procedure and stores data in sql cache and redis cache', async () => {
       spyOn(checkWindowDataService, 'sqlFindCalculationPeriodCheckWindow').and.returnValue({ id: 1, complete: false, checkEndDate: moment.utc().add(5, 'days') })
-      spyOn(schoolDataService, 'sqlFindSchoolIds').and.returnValue([1, 2, 3, 4])
+      spyOn(schoolScoresDataService, 'sqlFindSchoolIds').and.returnValue([1, 2, 3, 4])
       spyOn(pupilResultsDiagnosticCache, 'sqlDelete')
       spyOn(schoolScoresDataService, 'sqlExecuteGetSchoolScores').and.returnValue([{ id: 1 }])
       const sqlInsertSpy = spyOn(pupilResultsDiagnosticCache, 'sqlInsert')
@@ -27,7 +26,7 @@ describe('calculate-score-v2: v1', () => {
         fail()
       }
       expect(checkWindowDataService.sqlFindCalculationPeriodCheckWindow).toHaveBeenCalled()
-      expect(schoolDataService.sqlFindSchoolIds).toHaveBeenCalled()
+      expect(schoolScoresDataService.sqlFindSchoolIds).toHaveBeenCalled()
       expect(pupilResultsDiagnosticCache.sqlDelete).toHaveBeenCalled()
       expect(schoolScoresDataService.sqlExecuteGetSchoolScores).toHaveBeenCalled()
       expect(sqlInsertSpy.calls.first().args[1].generatedAt).toBeDefined()
@@ -39,7 +38,7 @@ describe('calculate-score-v2: v1', () => {
     })
     it('returns before proceeding further if no check window is found', async () => {
       spyOn(checkWindowDataService, 'sqlFindCalculationPeriodCheckWindow').and.returnValue({})
-      spyOn(schoolDataService, 'sqlFindSchoolIds')
+      spyOn(schoolScoresDataService, 'sqlFindSchoolIds')
       spyOn(pupilResultsDiagnosticCache, 'sqlDelete')
       spyOn(schoolScoresDataService, 'sqlExecuteGetSchoolScores')
       spyOn(pupilResultsDiagnosticCache, 'sqlInsert')
@@ -49,14 +48,14 @@ describe('calculate-score-v2: v1', () => {
         fail()
       }
       expect(checkWindowDataService.sqlFindCalculationPeriodCheckWindow).toHaveBeenCalled()
-      expect(schoolDataService.sqlFindSchoolIds).not.toHaveBeenCalled()
+      expect(schoolScoresDataService.sqlFindSchoolIds).not.toHaveBeenCalled()
       expect(pupilResultsDiagnosticCache.sqlDelete).not.toHaveBeenCalled()
       expect(schoolScoresDataService.sqlExecuteGetSchoolScores).not.toHaveBeenCalled()
       expect(pupilResultsDiagnosticCache.sqlInsert).not.toHaveBeenCalled()
     })
     it('returns before proceeding further if school ids are not found', async () => {
       spyOn(checkWindowDataService, 'sqlFindCalculationPeriodCheckWindow').and.returnValue({ id: 1, complete: false, checkEndDate: moment.utc().add(5, 'days') })
-      spyOn(schoolDataService, 'sqlFindSchoolIds')
+      spyOn(schoolScoresDataService, 'sqlFindSchoolIds')
       spyOn(pupilResultsDiagnosticCache, 'sqlDelete')
       spyOn(schoolScoresDataService, 'sqlExecuteGetSchoolScores')
       spyOn(pupilResultsDiagnosticCache, 'sqlInsert')
@@ -66,14 +65,14 @@ describe('calculate-score-v2: v1', () => {
         fail()
       }
       expect(checkWindowDataService.sqlFindCalculationPeriodCheckWindow).toHaveBeenCalled()
-      expect(schoolDataService.sqlFindSchoolIds).toHaveBeenCalled()
+      expect(schoolScoresDataService.sqlFindSchoolIds).toHaveBeenCalled()
       expect(pupilResultsDiagnosticCache.sqlDelete).not.toHaveBeenCalled()
       expect(schoolScoresDataService.sqlExecuteGetSchoolScores).not.toHaveBeenCalled()
       expect(pupilResultsDiagnosticCache.sqlInsert).not.toHaveBeenCalled()
     })
     it('returns before proceeding further if deleting the cache table fails', async () => {
       spyOn(checkWindowDataService, 'sqlFindCalculationPeriodCheckWindow').and.returnValue({ id: 1, complete: false, checkEndDate: moment.utc().add(5, 'days') })
-      spyOn(schoolDataService, 'sqlFindSchoolIds').and.returnValue([1, 2, 3, 4])
+      spyOn(schoolScoresDataService, 'sqlFindSchoolIds').and.returnValue([1, 2, 3, 4])
       spyOn(pupilResultsDiagnosticCache, 'sqlDelete').and.returnValue(Promise.reject(new Error('error')))
       spyOn(schoolScoresDataService, 'sqlExecuteGetSchoolScores')
       spyOn(pupilResultsDiagnosticCache, 'sqlInsert')
@@ -84,7 +83,7 @@ describe('calculate-score-v2: v1', () => {
         expect(error.message).toBe('error')
       }
       expect(checkWindowDataService.sqlFindCalculationPeriodCheckWindow).toHaveBeenCalled()
-      expect(schoolDataService.sqlFindSchoolIds).toHaveBeenCalled()
+      expect(schoolScoresDataService.sqlFindSchoolIds).toHaveBeenCalled()
       expect(pupilResultsDiagnosticCache.sqlDelete).toHaveBeenCalled()
       expect(schoolScoresDataService.sqlExecuteGetSchoolScores).not.toHaveBeenCalled()
       expect(pupilResultsDiagnosticCache.sqlInsert).not.toHaveBeenCalled()
